--- include/helpers/backend.hpp
@@ -1,11 +1,0 @@
-/**
- * Includes various helper functions for backend-focused functionality
- * such as random number generation
- */
-
-#ifndef BATTLESHIP_BACKEND_H
-#define BATTLESHIP_BACKEND_H
-
-int randomInt(int start, int end);
-
-#endif//BATTLESHIP_BACKEND_H
