--- src/helpers/backend.cpp
@@ -1,13 +1,0 @@
-/**
- *
- */
-
-#include "../../include/helpers/helpers.hpp"
-#include <random>
-
-int randomInt(int start, int end) {
-    static std::random_device rd; // Obtain a random number from hardware
-    static std::mt19937 eng(rd()); // Seed the generator
-    std::uniform_int_distribution<> dist(start, end);
-    return dist(eng);
-}
